Add data validation

+++ Controllers/CommentController.cs
@@ -2,6 +2,7 @@
 using api.Interfaces;
 using api.Mappers;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace api.Controllers;
 
@@ -21,14 +22,24 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var comments = await _commentRepo.GetAllAsync();
         var commentDto = comments.Select(s => s.ToCommentDto());
         return Ok(commentDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int} ")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var comment = await _commentRepo.GetByIdAsync(id);
         if (comment == null)
         {
@@ -41,6 +52,11 @@ public class CommentController : ControllerBase
     [HttpPost("{stockId:int}")]
     public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         if (!await _stockRepo.StockExists(stockId))
         {
             return BadRequest("Stock does not exist");
@@ -54,6 +70,11 @@ public class CommentController : ControllerBase
     [HttpPut("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto commentDto)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var comment = await _commentRepo.UpdateAsync(id, commentDto.ToCommentFromUpdateDto());
 
         if (comment == null)
@@ -67,13 +88,18 @@ public class CommentController : ControllerBase
     [HttpDelete("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var comment = await _commentRepo.DeleteAsync(id);
 
         if (comment == null)
         {
             return NotFound();
         }
-        
+
         return NoContent();
     }
 }
